+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - collector's method `remove` to clear observations with given label pairs [#263](https://github.com/tarantool/metrics/issues/263)
+- `counter:reset()` method [#260](https://github.com/tarantool/metrics/issues/260)
 
 ## [0.10.0] - 2021-08-03
 ### Changed